@@ -1,9 +1,57 @@
 module github.com/aaronland/go-picturebook-flickr
 
-go 1.12
+go 1.23.3
 
 require (
-	github.com/aaronland/go-picturebook v0.4.11
-	github.com/tidwall/gjson v1.6.8
-	gocloud.dev v0.22.0
+	github.com/aaronland/go-picturebook v0.9.3
+	github.com/tidwall/gjson v1.18.0
+	gocloud.dev v0.40.0
+)
+
+require (
+	github.com/MaxHalford/halfgone v0.0.0-20171017091812-482157b86ccb // indirect
+	github.com/aaronland/go-image v1.3.1 // indirect
+	github.com/aaronland/go-image-contour v1.0.2 // indirect
+	github.com/aaronland/go-image-halftone v0.1.0 // indirect
+	github.com/aaronland/go-mimetypes v0.0.2 // indirect
+	github.com/aaronland/go-roster v1.0.0 // indirect
+	github.com/aaronland/gocloud-blob v0.4.0 // indirect
+	github.com/dsoprea/go-exif/v3 v3.0.1 // indirect
+	github.com/dsoprea/go-iptc v0.0.0-20200609062250-162ae6b44feb // indirect
+	github.com/dsoprea/go-jpeg-image-structure/v2 v2.0.0-20221012074422-4f3f7e934102 // indirect
+	github.com/dsoprea/go-logging v0.0.0-20200710184922-b02d349568dd // indirect
+	github.com/dsoprea/go-photoshop-info-format v0.0.0-20200609050348-3db9b63b202c // indirect
+	github.com/dsoprea/go-utility/v2 v2.0.0-20221003172846-a3e1774ef349 // indirect
+	github.com/fogleman/colormap v0.0.0-20180829212827-f273ae61505a // indirect
+	github.com/fogleman/contourmap v0.0.0-20190814184649-9f61d36c4199 // indirect
+	github.com/fogleman/gg v1.3.0 // indirect
+	github.com/go-errors/errors v1.4.2 // indirect
+	github.com/go-pdf/fpdf v0.9.0 // indirect
+	github.com/go-xmlfmt/xmlfmt v0.0.0-20191208150333-d5b6f63a941b // indirect
+	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
+	github.com/golang/geo v0.0.0-20210211234256-740aa86cb551 // indirect
+	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/googleapis/gax-go/v2 v2.13.0 // indirect
+	github.com/mandykoh/go-parallel v0.1.0 // indirect
+	github.com/mandykoh/prism v0.35.3 // indirect
+	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646 // indirect
+	github.com/rwcarlsen/goexif v0.0.0-20190401172101-9e8deecbddbd // indirect
+	github.com/sfomuseum/go-exif-update v0.1.3 // indirect
+	github.com/sfomuseum/go-flags v0.10.0 // indirect
+	github.com/sfomuseum/go-font-ocra v0.0.3 // indirect
+	github.com/tidwall/match v1.1.1 // indirect
+	github.com/tidwall/pretty v1.2.0 // indirect
+	github.com/whosonfirst/go-ioutil v1.0.2 // indirect
+	go.opencensus.io v0.24.0 // indirect
+	golang.org/x/image v0.18.0 // indirect
+	golang.org/x/net v0.28.0 // indirect
+	golang.org/x/sys v0.24.0 // indirect
+	golang.org/x/text v0.17.0 // indirect
+	golang.org/x/xerrors v0.0.0-20240716161551-93cc26a95ae9 // indirect
+	google.golang.org/api v0.191.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240812133136-8ffd90a71988 // indirect
+	google.golang.org/grpc v1.65.0 // indirect
+	google.golang.org/protobuf v1.34.2 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
